Massively rename "strings" to "patterns".
Historically, we have used the term "string" for referring to each of the patterns defined in the "strings" section of a YARA rule. This was because the very first versions of YARA didn't support regular expressions or complex hex patterns, so they were basically plain sequences of bytes (with some wildcards at most). Later YARA adopted more complex patterns, but the "string" term stuck. In this new codebase I'm adopting the term "pattern" instead of "string", which is more accurate.

This also prevents confusing the type ast::String (now ast::Pattern) with the builtin Rust type String.