Use a Vector instead of a HashMap for storing namespaces and rules in the AST.
The reason for this change is that HaspMap iteration is not stable (items are iterated in a random order), but the compiler needs to iterate the rules in the order that they appear in the source code, in order to make sure that rules that are declared first appear first in the generated code.

Without this guarantee we may end up on a situation, where a rule `foo` that was declared after `bar`, and uses the result of `bar` in its condition, is executed *before* `bar`. If this happens the result of `bar` is not unknown by the time `foo` needs it.