fix: don't use binary search in `is_pat_match_at`
In general we can't rely on the matches being sorted by start offset. This assumption is correct for text patterns, but is not always correct for regular expression and hex patterns that may match strings of different lengths. In such cases, if the atom found by Aho-Corasick is in the middle or end of the pattern, the match start offset may be before the offset of a previously found match.