fix: remove the `uint64` and `uint64be` built-in functions
These two functions don't work when they return values that are larger than 0x7FFFFFFFFFFFFFFF as all integers are currently handles as signed, and therefore those values are interpreted as negative. In order to support `uint64` and `uint64be` we must also add full support for u64 integers.

This can be done at the WASM level, but requires some changes in `TypeValue`, and in the way the code is emitted for comparison operators. Particularly, the mix of signed and unsigned integers must be handled properly.