fix: return error when a global rule depends on a non-global rule
This is an error because global rules are evaluated before non-global rules, even if the global rule appears after the non-global one in the source code. This means that by the time the global rule is being evaluated we can't know if the non-global rule matched or not.

A global rule can depend on another global rule. And non-global rules can depend both on global rules and non-global ones.