refactor: improve the error reporting logic
Introduce the `SourceId` field on each `Span`, which indicates the original source code the span is associated with. With this change we don't need to pass a reference to the original source code to `ReportBuilder::build_report`, making the code simpler. Also, it allows to create multi-file error reports, showing code snippets from different source files. 

Also fixes an issue where duplicate rule declarations were not being detected if the rules where in different source files under the same namespace.