refactor: allow values in `TypeValue` to be constant or variable
Until now the value in a `TypeValue` could be only known or unknown by means of using `None` or `Some(T)`, but there was no way for indicating that a value was constant and therefore won't change at runtime, any value that was known at compile time was considered constant. This caused a bug in which global variables defined at compile time were being considered constant.