refactor: make some changes in the compiler API
Methods `add_source`, `define_global` and `new_namespace` now receives a mutable reference instead of taking ownership of the compiler object. 

The problem with the previous approach is that `add_source` is a fallible function. When the function succeeds it returns the compiler back to the caller, and the caller regain ownership, but when it fails the ownership is lost. This means that once `add_source` fails the compiler can't be used anymore, you can't add more sources or get the compiler rules.