refactor: redesign de WASM module builder
Putting the code for all conditions in a single WASM function didn't scale well. the `wasmtime` module has a limit in the size of the functions it can accept. Also, by using a single function we couldn't benefit from the parallel code generation offered by `wasmtime`.

With this change hundreds of thousands of rules can be compiled, and the compilation time gets reduced.