fix: issue with negative numbers in operators `at` and `in`
The compiler returns an error on expressions like `$a in (-1..4)` or `$a at -1` because match offsets must be positive. However, at runtime this operations can receive negative offsets if they come from operations that the compiler can't evaluate at compile time. The runtime must be prepared for handling such cases correctly.