fix: issue with very large timeouts
The default scanner timeout is 315.360.000 seconds instead of  `u64::MAX`. Using `u64::MAX` as the default value doesn't work because that value is later added to the current epoch and would cause an integer overflow. We need a value that is large enough but that has some room before `u64::MAX` is reached.