perf: allow anchored patterns with offsets != 0
With this change any pattern that is used exclusively in a expression like `$a at <offset>` is not included in the Aho-Corasick automata because the offset where it can match is known beforehand. This was already done for patterns at offset 0, but now the mechanism is generalized for any offset.