refactor: avoid leaving WASM code debris when `add_source` fails
With this change the emission of WASM code is the last step in the compilation process, after all fallible functions have been executed successfully. This way we prevent leaving WASM code debris when the compilation fails. In order to implement this change it was necessary to split the `Context` type into two new types `CompileContext` and `EmitContext`.