feat: implement the `lnk` module (#45)
The lnk module parses Windows LNK files and exposes some of their metadata to YARA rules. This module is inspired by @BitsOfBinary's implementation of a similar module YARA in C, but it's greatly simplified and only exposes a subset of the metadata.

More enhancements can be made in the future.