fix: prevent conflict between types defined by different YARA modules.
This commit fixes an issue that prevented two different modules from defining structures with the same name. For example, if both `module_1.proto` and `module_2.proto` defined a structure as  `message Foo {...}`, both structures were colliding.

With this change you can use the `package` statement in the the `.proto` files that defines a YARA module, which prevent these kind of collisions.