fix: improve `rva_to_offset` so that it works in certain edge cases.
With this fix `rva_to_offset` handles certain cases where the raw data size of a section is larger than its virtual size. Before this, if the virtual size was not zero, it was used as the section size while determining if an RVA belonged to the section or not. However, in cases were the raw data size is larger, some RVA were not properly translated into a file offset.

An example of such case is file db6a9934570fa98a93a979e7e0e218e0c9710e5a787b18c6948f2eedd9338984, where import information was not properly extracted because of a failure while translating the RVA of the imports section to a file offset.