fix: issue when modules and global variables were used at the same time.
The root is of the issue was that global variables and modules were kept in separate structures during compile time, but they are together in the same struct at scan time. This caused that field indexes where computed with respect to either one structure or the other while compiling the rules, but at scan time these field numbers were not valid.