feat: YAML serializer (#53)
* feat: draft an implementation of a YAML serializer for protobufs.

This introduces the `yara-x-proto-yaml` crate, which has a minimalistic API that takes a protobuf message and produces a YAML representation of it.

* fix: make the order of map entries stable

The `map` type in a protobuf is internally represented by a `HashMap`, which doesn't guarantee that items are always iterated in the the same order. The order may vary between executions, even if the map's content is the same. This means that we are forced to sort the items in the map before serializing them to YAML.

* feat: protobuf to yaml serialization

* fix: errors that occured while merging

* fix: clippy warnings

* chore: update comments

* feat: add support for PE module output in dumper

---------

Co-authored-by: Victor M. Alvarez <vmalvarez@virustotal.com>