tests: use YAML in golden files
The YAML format is better than the protobuf text format for golden files because field order in maps is stable. The protobuf text format doesn't guarantee any order in map fields, because it relies directly on the underlying protobuf structr, which uses a HashMap for storing maps.

In the other hand, the YAML serializer always sort maps by keys before producing the YAML output.