perf: improve performance by grouping multiple lookup operations
When performing field lookup operations, we need to call Rust code from WASM. As these calls are expensive we need to minimize the number of calls as much as possible. When an expression like `foo.bar.baz` are evaluated, instead of emitting three lookup operations (one that finds `foo` in the root structure, one that finds `bar` in the `foo` structure, and finally one that finds `baz` in the `bar` structure) we only emit one lookup operation that actually does the three in a single call to Rust code.