fix: issue in error reports when the source file contains UTF-8 characters longer than a byte.
The root cause of this issue is that the `ariadne` crates works with character-wise spans, while our AST works with byte-wise spans. Our AST gets the spans from `pest`, so switching to character-wise spans is not option. Also, I think that byte-wise spans are easier to work with because they can be used directly with `&str` types containing the source code.

In the future we may consider adapting `ariadne` to our needs by making it work with byte-wise spans, but after some examination of the source code it doesn't look like a trivial task.