feat: implement method calls (#59)
A mechanism for supporting method calls is now implemented. Only structs can have methods at this time, but in the future this can be extended to other types like maps and arrays.

The `module_export` and `wasm_export` now have a new argument `method_of` that allows to specify that a WASM-exported function is a method of some type. For example:

```
#[module_export(method_of = "test_proto2.NestedProto2")]
fn some_function(ctx: &mut ScanContext, structure: Rc<Struct>) {
  ... // here the `structure` argument is a `NestedProto2`
}
```