perf: some optimization in `math` module
As `distribution[i]` contains the number of occurrences of byte `i` in the data, and we have 256 entries in `distribution` (one per possible byte), the sum of all entries in `distribution` is equal to the number of bytes in the data. Instead of computing the sum of all `distribution[i]`, use `data.len()` which is faster.