perf: improve the implementation of `split_at_large_gaps`
This function splits a pattern into multiple chained patterns if they contain very large gaps. The previous implementation splits the pattern even if the resulting patterns are very short. For instance, `{01 02 [0-3000] 03}` was split into `{01 02}` and `{03}`.

The problem with this is that `{03}` is very short, and having very short atoms is problematic. The new implementation guarantees that all patterns resulting from the split match at least 2 bytes.