chore: rename top-level directories.
In the future we want to add a `go` directory that contains the Go bindings for YARA-X. Having a `yara-x-go` directory is not ideal, because the import path for that Go module would be `github.com/VirusTotal/yara-x/yara-x-go`. Instead we want the more natural `github.com/VirusTotal/yara-x/go`. However, adding a `go` directory makes the current naming incoherent, but with this change everything fits well again.