refactor: replace the `YRX_MOD_PROTO_DIRS` env variable with `YRX_PROTOC_CONFIG_FILE`
Instead of having an environment variable that points to a directory with `.proto` files, now the environment variable points to a JSON file, that in turns contains the list of `.proto` files that should be included in the compilation.

This way we can control which individual `.proto` files are included, in a directory that could contain other unrelated `.proto` files that don't need to be compiled.