refactor: change the logic used for specifying additional modules (#83)
This removes the use of the YRX_PROTOC_CONFIG_FILE environment variable, in favor of two new environment variables YRX_EXTRA_PROTOS and YRX_EXTRA_PROTOS_BASE_PATH. The previous mechanism for specifying extra modules was more complex and didn't fit well with Bazel which was the primary use case for this mechanism.

The new YRX_EXTRA_PROTOS accepts a list of space-separated file paths pointing to .proto files. This paths are either absolute or relative to the build.rs file. With YRX_EXTRA_PROTOS_BASE_PATH you can make the paths relative to some other point in the file system.