fix: remove the logic that parses signatures from Macho files.
This code depended on the `cryptographic-message-syntax` crate, which introduces a ridiculous amount of indirect dependencies, including crates like `tokio` and `reqwest`. I don't know why a crate devoted to parsing cryptographic data structures depends on a HTTP client library.

I've tried replacing `cryptographic-message-syntax` with `cms`, but it doesn't seems to be a straightforward change. We need to find a way to implement this feature using `cms` or some other crate that doesn't introduce insane dependencies.