fix: be more flexible about the number of PE sections parsed
Before this change the parser was strict about the number of PE sections correctly parsed. If the parser was not able to parse as many sections as the number of sections indicated in the header, the parser failed.

Now it tries to parse the number of sections indicated by the header, but as soon as it is able to parse at least one section, it succeeds.