feat: interpret the sequences `\<` and `\>` in regexp as literals instead of word boundaries
The `regex_syntax` crate interprets `\<` and `\>` as start of word and end of word boundaries respectively. However, in YARA this has always been interpreted as the escaped form of `<` and `>` literals. In order to keep backward compatibility, this commits introduce the `Transformer` type, which receives the AST produced by `regex_syntax` and replace the word boundary tokens by literal ones.