tests: fix misleading test
An existing test leaded into thinking that the dot (.) in regular expressions matched unicode characters by default, but that's not the case. The dot matches a single byte, unless unicode mode is explicitly enabled by using the `(?u)`. In that case the dot matches an unicode character, no matter if it takes more than one byte.