fix: issue while parsing ELF files with incorrect size in section headers
File 71adb87ee8ee76f32f54c70584ef14f67a4bc6f55df3f847c344726405927a1e has an incorrect size in the section header corresponding to the section that contains the string table. As we strictly applying that limit, the string table was truncated, even though it is perfectly valid in the file.

YARA's parser was more relaxed and didn't take into account the size indicated in the section header, therefore was able to get strings from the string table without issues. With this change we follow the same strategy in YARA-X.