fix: issue while building the crate by `doc.rs`
When the crate is built by `doc.rs` the source code is put in a read-only file system and we can't modify source files. As we were doing with `modules.rs`. Now `build.rs` detects when the create is being built by `doc.rs` and doesn't update the `modules.rs` file. The `add_modules.rs` file is now stored in the repository for consistency and simplicity.