feat: Added support for non-canonically DER encoded (counter)signatures (#99)
This adds deferred parsing of SignedAttributes field of SignedInfo
structure within CMS, allowing us to directly access the raw data of
this field so that it can be properly verified without canonical DER
decoder in the way.

To make things easier and avoid rewrite of as many code as possible,
there's option to turn `DeferSignerInfo` structure directly into
`SignerInfo` structure. On top of that, all other fields are decoded
as usual, making deferred structure still useful in cases where
non-deferred attributes are needed.

Affected binary added to tests.