feat: implement Authenticode parsing and verification without relying on OpenSSL (#100)
Until now we were using the authenticode-parser crate for Authenticode parsing and verification. This is simply a Rust wrapper around https://github.com/avast/authenticode-parser which is written in C and uses OpenSSL under the hood. Depending on OpenSSL makes building and deploying YARA-X harder, specially when you want to integrate YARA-X in other systems.

With this change all the Authenticode parsing and validation is re-written in Rust.

---------

Co-authored-by: Marek Milkovič <milkovic.marek@gmail.com>