fix: don't abort PE parsing when `opt_hdr.magic` is incorrect.
When a file is corrupt `opt_hdr.magic` may have an incorrect value that doesn't correspond to a 32-bit nor a 64-bit file. In such cases, instead of making the parsing fail, try to continue as it was a 32-bit file even if some values are incorrect. This is what YARA does, and this makes YARA-X closer to how YARA works in these edge cases.