feat: implement an option for allowing invalid escape sequences in regular expressions
Historically, YARA has accepted any character that is preceded by a backslash in a regular expression, even if the sequence is not a valid one. For instance, `\n`, `\t` and `\w` are valid escape sequences in a regexp, but `\N`, `\T` and `\j` are not. However, YARA accepts all of these sequences. The valid escape sequences are interpreted as their special meaning (`\n` is a new-line, `\w` is a word character, etc.), while invalid escape sequences are interpreted simply as the character that appears after the backslash. So, `\N` becomes `N`, and `\j` becomes `j`.

This change introduces the `Compiler::relaxed_regexp_escape_sequences` API, which allows to turn on an option that makes YARA-X to behave in the same way than YARA with respect to invalid escape sequences in regular expressions. This option is turned off by default.

Also, the option `--relaxed-escape-sequences` is added to the CLI.