feat: generalize the `relaxed_escape_sequences` option to `relaxed_re_syntax`
With this change, the new option not only controls whether invalid escape sequences should be accepted, it also relaxes other cases in which the YARA-X is more strict with regular expressions than YARA. This is the case with characters that have a special meaning in a regular expression, but that YARA treats as literal if they appear in a context where this special meaning doesn't make sense.

For instance, YARA interprets the curly braces in `/foo{}bar/` as literal characters, while in `/foo{1,2}bar/` they are interpreted as part of the repetition operator `{1,2}`.  YARA-X doesn't accept `/foo{}bar/` as a valid regular expression, unless you enable the relaxed_re_syntax option.