perf(py): avoid creating new tuples when not necessary.
Before this change, every time you called a getter that returns a tuple (like Pattern.matches, Rule.patterns, Rule.metadata) the tuple was built again from scratch. This is not necessary, as those tuples are immutable. Now the tuples are created once, and a reference to them is returned every time you need to get them.