refactor(cli): change the way in which the output path is passed to the `compile` command
The output path was passed as the last positional argument for the command, but this have an inconvenience. If the user pass multiple input files and forgets to specify an output path, the last input path is used as output and the input file is overwritten.

In order to prevent this, the output path is now passed with the `--output` option. If the option is missing, the default path `output.yarc` is used.

Closes #126.