refactor: start working on a new fault-tolerant parser (#152)
The existing parser, based in the pest crate, has been good-enough so far, but lacks some features that makes it unsuitable for some tasks, like implementing a language server for YARA that can be integrated in text editors like VSCode.

For implementing a language server we need a parser that is fault-tolerant, recovering from syntax errors as much as possible and parsing the code that comes after the problematic code. As an additional benefit the CLI tool can report multiple syntax errors at once, instead of a single one as it currently does.

The new parser should be faster too.