feat: implement new PEG parser (#154)
The current parser, based on the `pest` crate, has been adequate as a starting point but lacks several features that make it unsuitable for specific tasks, such as implementing a language server for YARA that can be integrated into text editors like VSCode.

To implement a language server, we need an error-tolerant parser capable of recovering from syntax errors and continuing to parse the subsequent code. Additionally, this enhancement allows the CLI tool to report multiple syntax errors simultaneously, instead of just one as it currently does.

The new parser is a hand-written PEG parser.