refactor: check pattern modifiers in the AST->IR phase
Instead of checking for duplicate pattern modifiers and invalid modifier combinations in the CST->AST phase, we now do it in the AST->IR phase. The idea is that the CST->AST phase should perform as many checks as possible. Most semantic validations will be done in the AST->IR phase.