fix: Fix handling invalid(?) unicode in filenames.
When running yr scan against some random files I have laying around I noticed a
crash. I think it is because there is invalid (at least I think it is, I'm so
bad at understanding unicode) unicode in the filename of one of the files
contained in 84523ddad722e205e2d52eedfb682026928b63f919a7bf1ce6f1ad4180d0f507
(available on VT).

This changes it so the string is using the debug formatter which handles the
invalid utf-8 better? This seems to work fine on my machine.