refactor(go): change the result of `Scanner.Scan` to `ScanResults`
`Scanner.Scan` was returning a slice with the matching rules, but this is not flexible enough if we want the results to include additional information like non-matching rules or module outputs. The `ScanResults` object provides access to the matching rules and to any other information that we want to return in the future.