Add "modules" subcommand to "debug".
Add the ability to print available modules via "yr debug modules".

I experimented with making it take a "verbose" flag that would also print the
protobuf definition of the module, which would make it easier for users to find
out what is "available" to use from the module in their rules, but it was a lot
to print out and would require users to translate from the protobuf types into
YARA types, which is probably more than they want to do.