feat: add `modules` subcommand to display modules available. (#179)
* Add the ability to print available modules via `yr debug modules`.

I experimented with making it take a `verbose` flag that would also print the
protobuf definition of the module, which would make it easier for users to find
out what is "available" to use from the module in their rules, but it was a lot
to print out and would require users to translate from the protobuf types into
YARA types, which is probably more than they want to do.

* Add `debug-cmd` feature and put `debug` command behind it.

Add a `debug-cmd` feature that enables the `debug` command. Prior to this the
command was always available but just hidden. This makes it a compile-time
feature that can be useful if you're into doing developer type things like
debugging the syntax trees.