ci: keep only `cp39` in the build matrix for the Python extension.
With `cp39` we are generating wheel files that are compatible with Python 3.8 and above. The wheel file built by `cp310`, `cp311` and `cp312` are the same as the one built by `cp39`. This is problematic when uploading the files to pypi.org, because trying to upload files with the same name often result in file corruption.

See: https://github.com/actions/download-artifact/issues/298