chore: prevent re-compiling due to file changes caused by `build.rs`
`build.rs` updates `src/modules/add_modules.rs` and `src/modules/modules.rs` every time it runs, but the same time `cargo:rerun-if-changed=src/modules` tells `cargo` to compile the crate again if a change occurs in `src/modules`. This mean that the `build.rs` script is invalidating the compiler's all the time.

Closes #187