feat: allow controlling the access to certain fields in a module using compiler features.
This introduces a `Compiler::enable_feature` method that allows enabling features with arbitrary names. These features can be used in the `.proto` file that defines a YARA-X module for controlling whether a structure field is accesible or not. You can force a field to become accessible only when one more features has been enabled in the compiler.

For the time being this only works when `.proto` files are compiled using the official Protobuf compiler (i.e: the `cargo --features=protoc`)