feat(cli): multiple changes in `fmt` command.
Added the `--check` option for running `fmt` in check mode. In this mode the command doesn't modify the files, only returns exit code 0 if the formatting is correct and 1 if the file needed some change. The `--write` option is removed, by default the command applies the changes to the input files.