feat: expose an API for converting the `potentially_slow_loop` warning into an error.
Also improve the logic for detecting potentially slow loops. In addition to the use of `filesize` as the upper bound of loop ranges, it also looks for the use the number of occurrences of some pattern (i.e: `for any x in (0..#a)`), but doesn't raise the warning when the `math.min` function is used for limiting the range's upper bound.