feat: serialize compiled rules in a platform-independent way. (#202)
With this change the compile command produces rules that can be used across different platforms. You can compile a set of rules in a Linux and use them in a Windows. Until now, the compiled rules were platform-specific, and could be used only in the same platform where it was compiled.