Implement toml config for fmt command.
This implements a config option which can be used to set options for "yr fmt".

An example config file is included as "wxsfmt.toml" which currently has
everything disabled just to demonstrate the behavior.

If any option is not set it will fall back to the default values provided by the
"fmt" command which currently match the defaults provided by the library.