Improve config file handling.
Move the config structures and loading to a separate module, so that it can be
used by each command.

Pass the "fmt" section of the config to the "fmt" command. As we add support for
other commands to use the config file we can easily pass their sub-structure
into the command.

Load the config file from %{HOME}/.yara-x.toml - or whatever the corresponding
location is on Windows.

The exact format of the config is still to be documented, but now that it is
generic and not specific to formatting I've moved the docs to a new file that is
more appropriately named.