feat: Add support for newlines before curly braces.
This commit adds support for newlines before curly braces in rule declaration.
If "true" it will ensure at most 2 newlines exist before the opening curly
brace. If zero newlines exist it will insert only one, but if 2 or more exist it
will reduce it to exactly 2.

The default for this is "false".