feat: more `fmt` improvements (#224)
This adds support for optional newline before curly brace in the rule declaration.

```
// This is the default
rule a {
```
vs

```
rule a
{
```

It also adds support for empty lines before and after section headers. Each one is a separate option - you can have them before section headers but not after, vice versa, both, or none.

```
rule a {

  meta:

    foo = "bar"
```

vs

```
// This is the default (no empty lines)
rule a {
  meta:
    foo = "bar"
```

This also fixes an issue where we were not inserting newlines before meta definitions, so if you had two meta definitions on the same line the formatter left them like that. They are now fixed so that each meta definition goes on a separate line.