add tests for fmt config options. (#230)
This commit adds tests for the configurable formatter. Each configuration option
is tested individually, as they each should be independent of each other. If we
have a need to test multiple configuration options simultaneously in the future
this setup should be able to do it easily.

While here I also found a bug in how tab stops were treated by the formatter
when removing trailing whitespaces. After indentation was performed the step to
remove trailing whitespace was not properly checking for the inserted tab
characters.

There was also another bug with handling the emptyline before the section header
which I fixed.

Also, fix a bug in the docs where the `meta.align_values` and
`meta.align_patterns` options were incorrectly set. They changed to true in a
previous commit but I did not update the docs then.