fix: issue in CSE algorithm causing that portions of the IR tree was not being traversed.
During the DFS traversal of the IR, it was pruned with the intention of avoiding traversing the descendants of some node, but this was done in the "leave" event, causing that the siblings of the node were not traversed.

To avoid traversing the descendants of a node without avoiding the siblings, the call to `prune` must be done during the "enter" event.