feat: Implement extra data from dump mode.
This commit is a prototype for adding a "dump" mode for modules. This is a flag
that can be passed to modules that tells them to perform "extra" processing that
they would normally not do during "scan" mode. This is meant for things like
`pe.imphash()` which is normally calculated on-demand.