Implement required metadata.
This implements a basic linter into the compiler that can check for required
metadata fields and types. I chose to implement it into the compiler so that we
can get these checks for free across any tool/binding that uses the compiler
directly. Also, it was easier to get nicer reporting by building it into the
compiler.

This is an optional feature that you can turn on in the compiler using the
"required_metadata" method on the compiler. If you do this the compiler will
issue warnings if any metadata is missing or if the type is incorrect.

I've provided all the basic types (string, int, bool, float) and also some
convenience types like md5, sha1, sha256 and hash to make it a bit more flexible
for users. Extending these with more convenience types is an easy exercise too.