Rework config file handling.
Prior to this change if you had an invalid config file (ie: using the wrong
type for an option) we would just silently replace it with the default config,
which made debugging way more difficult than it should have been.  We now pass
errors back up to the caller if the error is a because the config file can not
be parsed.

Errors now look like this:

```
Error: invalid type: found signed int `0`, expected a boolean for key "default.fmt.patterns.align_values" in .yara-x.toml TOML file
```

Also, because we now display errors like the above it doesn't make sense to have
each command have a -C option. If the config file in ${HOME}/.yara-x.toml is
improperly formatted we would try to parse it and display an error even before
we called the subcommand. This meant if you had an invalid config file in the
default location and used "yr fmt -C /tmp/foo ..." you would see the error from
your default config file despite the -C argument.

To deal with this the -C argument was removed from each command and moved into
the main one. It now loads the default config if, and only if, you do not
specify one with the -C argument.