Workaround bug in globwalk.
As identified in #280 there is a bug in globwalk when using relative paths. Fix
it by trying to canonicalize the path, and if that fails falls back to using the
relative path. This is fine because the reason canonicalization failed is
because the relative path was to a place that doesn't exist, and that does not
trigger the bug (it errors out before we get there). We intentionally want this
behavior so we get the nicer error reporting with colors. ;)

Fixes #280.