feat: allow using previously defined variables within a `with` statement.
Now the variables declared in a `with` expression can be used by expressions that appear later within the same `with` statement. For instance, this is now valid:

```
with
  foo = <expr>,
  bar = foo + <expr>: (
    . . .
)
```

Notice how the variable `foo` can be used in the expression that defines the value of `bar`. This was not possible before.

Closes #287