Don't strip prefix if path is exactly "."
One of the things I often run is "yr scan -r rules.yara ." when I have a
collection of samples and I want to scan all of them in the current directory.

With the fix in #280 the path was being stripped to an empty string which would
result in an error when trying to scan. Fix it by only performing the strip if
the path is not equal to ".".