fix: prevent the `fmt` command from modifying non-YARA files.
The `fmt` command didn't take into account if the input file was actually a YARA source code file or not, it tried to format the file anyways, modifying binary files or whatever you passed to it. Now the `fmt` fails if the input file contains invalid UTF-8 sequences, and don't modify the files in such cases.

Closes #286.