refactor: introduce the `Compiler::add_linter` API.
Instead of adding multiple new methods to the `Compiler` object for each type of check, add a single `add_linter` method that takes a type that implements the `Linter` trait. Each check is implemented as a new type that implements this type and put under `yara_x::linters`.

With this change the compiler's API is not bloated with multiple methods and we can introduce more linters in the future with more complicated APIs.