Add tag linter.
Add a tag linter that can be configured to use either a list of allowed strings,
or a regexp. If both are provided in the config file the allowed list is
preferred by the "check" command as it is more explicit.

As part of this work I realized we have no way to return multiple linter
findings in a single check call, which is necessary when checking tags, so I
modified the internal API to return a Option<Vec<LinterResult> where no linter
findings is represented by returning None.

I also discovered that the way the compiler works is that any linter errors are
collected but we only return the first one when calling "add_source()" which
is not ideal when you can have multiple linter errors collected up. To deal with
this I added a "linter_errors: Vec<CompileError>" to the compiler struct so that
they can be checked for separately from actual compiler errors.

I also had to modify the output handling of the "check" command so that multiple
linter errors are handled correctly.

Lastly, a couple of typos and copy/paste mistakes in documentation were fixed.